lvol: use uint64_t to represent lvol sizes
Replace the few existing uses of size_t as lvol size with uint64_t for
consistency.  size_t is meant to represent the size of an object in
memory, and it may be smaller than uint64_t (e.g. on 32-bit platforms).

+++ lib/bdev/lvol/vbdev_lvol.h
@@ -52,7 +52,7 @@ int vbdev_lvs_create(struct spdk_bdev *base_bdev, const char *name, uint32_t clu
 void vbdev_lvs_destruct(struct spdk_lvol_store *lvs, spdk_lvs_op_complete cb_fn, void *cb_arg);
 void vbdev_lvs_unload(struct spdk_lvol_store *lvs, spdk_lvs_op_complete cb_fn, void *cb_arg);
 
-int vbdev_lvol_create(struct spdk_lvol_store *lvs, const char *name, size_t sz,
+int vbdev_lvol_create(struct spdk_lvol_store *lvs, const char *name, uint64_t sz,
 		      bool thin_provisioned, spdk_lvol_op_with_handle_complete cb_fn, void *cb_arg);
 
 void vbdev_lvol_create_snapshot(struct spdk_lvol *lvol, const char *snapshot_name,
@@ -69,7 +69,7 @@ void vbdev_lvol_create_clone(struct spdk_lvol *lvol, const char *clone_name,
  * \param cb_arg Completion callback custom arguments
  * \return error
  */
-void vbdev_lvol_resize(struct spdk_lvol *lvol, size_t sz, spdk_lvol_op_complete cb_fn,
+void vbdev_lvol_resize(struct spdk_lvol *lvol, uint64_t sz, spdk_lvol_op_complete cb_fn,
 		       void *cb_arg);
 
 void vbdev_lvol_rename(struct spdk_lvol *lvol, const char *new_lvol_name,
